Revert "Fix: Aggressively transpile dependencies to resolve build error"
This reverts commit 618df058e483122df9970e1ed9ad932d735660c7.
