+++ .env.example
@@ -15,10 +15,4 @@ TEST_THERMOSTAT_UUID=5cdad4517dda
 DEBUG=false
 
 # The URL to forward thermostat cloud data to (optional).
-FWD_URL=
-
-# Service token for scanner (generate this as described earlier)
-SERVICE_TOKEN=your_service_token_for_scanner
-
-# (Optional) Secret for service token endpoint, if you use one
-SERVICE_TOKEN_SECRET=your_service_token_secret
+# FWD_URL=http://some_other_server/
