+++ .gitignore
@@ -23,3 +23,9 @@ build/
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# database
+thermostat_data.db
+
+# logs
+server.log
