+++ context/HostnameContext.js
@@ -1,19 +1,18 @@
-import React, { createContext, useState, useEffect } from "react";
+import React, { createContext, useState } from "react";
 import { Platform } from "react-native";
 
-export const HostnameContext = createContext("Loading..."); // ✅ Default value
+export const HostnameContext = createContext(null);
 
 export const HostnameProvider = ({ children }) => {
-  const [hostname, setHostname] = useState(null);
+  // For web, use localhost:5000 for development.
+  // For production, this should be replaced with the actual server URL.
+  // For mobile, you'll need to use the IP address of your development machine.
+  const backendUrl = Platform.select({
+    web: 'http://localhost:5000',
+    default: 'http://<your-dev-machine-ip>:5000', // Replace with your IP
+  });
 
-  useEffect(() => {
-    const fetchHostname = async () => {
-      let host = Platform.OS === "web" ? window.location.hostname : "Unknown";
-      setHostname(host);
-    };
-
-    fetchHostname();
-  }, [hostname]);
+  const [hostname, setHostname] = useState(backendUrl);
 
   return (
     <HostnameContext.Provider value={hostname}>
