+++ package.json
@@ -2,38 +2,39 @@
     "main": "index.web.js",
     "name": "radio_thermostat_project",
     "version": "0.0.1",
-    "proxy": "http://192.168.0.21",
     "private": true,
     "scripts": {
         "start": "npx expo start",
         "web": "npx expo start --web",
-        "shart": "node node_modules/react-native/local-cli/cli.js start"
+        "shart": "node node_modules/react-native/local-cli/cli.js start",
+        "server": "node server/server.js"
     },
     "dependencies": {
         "@babel/runtime": "^7.27.0",
         "@expo/metro-runtime": "~4.0.1",
-        "@react-native-async-storage/async-storage": "^2.1.2",
-        "@react-native-community/slider": "^4.5.6",
+        "@react-native-async-storage/async-storage": "1.23.1",
+        "@react-native-community/slider": "4.5.5",
         "axios": "^1.9.0",
         "bcryptjs": "^3.0.2",
         "better-sqlite3": "^11.10.0",
         "body-parser": "^2.2.0",
         "cors": "^2.8.5",
         "expo": "~52.0.46",
-        "expo-sharing": "^13.1.5",
+        "expo-sharing": "~13.0.1",
         "express": "^5.1.0",
         "js-cookie": "^3.0.5",
         "jsonwebtoken": "^9.0.2",
         "jwt-decode": "^4.0.0",
         "metro-runtime": "^0.82.1",
         "net-snmp": "^3.22.0",
+        "node-arp": "^1.0.6",
         "react": "18.3.1",
         "react-dom": "18.3.1",
         "react-native": "0.76.9",
         "react-native-chart-kit": "^6.12.0",
         "react-native-gesture-handler": "~2.20.2",
         "react-native-network-info": "^5.2.1",
-        "react-native-svg": "^15.11.2",
+        "react-native-svg": "15.8.0",
         "react-native-vector-icons": "^10.2.0",
         "react-native-web": "~0.19.13",
         "react-native-webview": "13.12.5",
@@ -43,6 +44,7 @@
     "devDependencies": {
         "@babel/plugin-transform-private-methods": "^7.25.9",
         "@react-native-community/cli": "latest",
+        "babel-loader": "^8.2.3",
         "metro-config": "^0.82.1",
         "metro-react-native-babel-preset": "^0.77.0"
     }
